Made significant changes to the internal representation of rigidbodies
Changed the internal representation of rigidbodies to include the Box2d and Bullet rigidbody representations. Made major changes to the internal hypothetical structure of engine::physics::physobject.

Changed the include/graphics/vkGlobals variables to be extern.

Fixed __STRINGIFY().

+++ include/physics/physobject.hpp
@@ -79,13 +79,13 @@ namespace engine
 				*	@param x X position
 				*	@param y Y position
 				*/
-				void setPos(int x, int y);
+				void setPos(double x, double y);
 				/*	Set the position
 				*	@param x X position
 				*	@param y Y position
 				*	@param z Z position
 				*/
-				void setPos(int x, int y, int z);
+				void setPos(double x, double y, double z);
 				/*	Set the position
 				*	@param pos The position
 				*/
@@ -99,13 +99,13 @@ namespace engine
 				*	@param x Pitch
 				*	@param y Yaw
 				*/
-				void setRot(int x, int y);
+				void setRot(double x, double y);
 				/* Set the rotation
 				*	@param x Pitch
 				*	@param y Yaw
 				*	@param z Roll
 				*/
-				void setRot(int x, int y, int z);
+				void setRot(double x, double y, double z);
 				/*	Set the rotation
 				*	@param pos The rotation
 				*/
@@ -119,13 +119,13 @@ namespace engine
 				*	@param x X scale
 				*	@param y Y scale
 				*/
-				void setScale(int x, int y);
+				void setScale(double x, double y);
 				/* Set the scale
 				*	@param x X scale
 				*	@param y Y scale
 				*	@param z Z scale
 				*/
-				void setScale(int x, int y, int z);
+				void setScale(double x, double y, double z);
 				/*	Set the scale
 				*	@param scale The scale
 				*/
@@ -138,7 +138,7 @@ namespace engine
 				// Advanced physics stuff
 
 				/*	Sets the owner of the physobject.
-				*	Assumed to be called when being initialized into the scene.
+				*	Assumed to be called when being initialized doubleo the scene.
 				*	Locks so it can't be called twice.
 				*	@warning FOR INTERNAL OR EXPERT USE ONLY, CAN CAUSE SEVERE ISSUES
 				*	@param owner Pointer to the owner, assumed to be either an object, or a simulation for internal use.
